# becker/lock-mq-msg