Dearest Reviewer,

A possible fix for issue #23. If the mq_poll was called within the same
microsecond it is possible they both return the same results. Locking
the mq_msg for update holds them for the `BEGIN...END` of the function
return. I did notice a slight slowdown on my branches vs main but I
think its better to not reprocess a job.

This branch:
```stress
min: 0.068051966s
max: 41.474278327s
median: 32.28066282s
95th percentile: 39.407505295s
throughput: 239.26637244549013/s
```

Main:
```stress
min: 0.038561309s
max: 37.321387572s
median: 29.594086093s
95th percentile: 34.564979967s
throughput: 263.7400277852498/s
```

Thanks
Becker

Ps: If reading issue 23 my coworker has confirmed this change does prevent
the dup processing for them.

It is hard to tell but its adding a FOR UPDATE in the SELECT * FROM mq_msgs
https://github.com/StructionSite/sqlxmq/pull/1/files#diff-0caca38de8ed18e06509d8461abe4f9a4dac0a3075dba8976ae68d90ee255b7aR31