@sbeckeriv 2022-02-08
author addressed it with a rewrite of the sql with no lock.